Project2_Code/requirements.txt
absl-py==0.9.0
antlr4-python3-runtime==4.8
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base==1.0.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
doqu==0.28.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2
matplotlib==3.1.3
numpy==1.18.1
oauthlib==3.1.0
onnx==1.6.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
props==0.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
reference==0.0.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
svgwrite==1.3.1
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminaltables==3.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
traitlets==4.3.3
Tree==0.2.4
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
